Minor candlestick json converter backwards compatility fix (#7428)
- Minor bug fix for candlestick json converter backwards compatibility,
  detected running the new report generator on an old lean backtest
  result json

+++ Common/ChartPoint.cs
@@ -82,7 +82,7 @@ namespace QuantConnect
         /// Shortcut for <see cref="x"/> for C# naming conventions
         /// </summary>
         [JsonIgnore]
-        public decimal X => x;
+        public long X => x;
 
         /// <summary>
         /// Shortcut for <see cref="y"/> for C# naming conventions
